refactor(editor): extract useEditorImage (Phase 1a)
First step of the EditorOverlay decomposition. Move the image load lifecycle
(decode → downscale-on-failure → error) + the full-res sourceImage / capped
previewSource / sourceImageRef into useEditorImage. EditorOverlay keeps the
"open a new image" reset + cached-depth lookup, and consumes the hook's values;
the Apply/Text-apply promote sites use the exposed setters + ref unchanged.

Behavior-preserving: all 25 editor/golden e2e specs green (03/06/07 + 18/19/20).
